redis: merge four files into two
redis_signal_dash_zeroconf into redis_signals, and
redis_signal_handle_config_request into redis_signal_handle_config.

None of these four was big enough to warrant a separate file, and
they're very close in function.
